requirements_org.txt
asgiref==3.5.2
Django==4.1.3
django-debug-toolbar==3.7.0
django-extensions==3.2.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.2
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
